requirements.txt
jinja2
rjsmin
rcssmin
numpy
lxml
orjson
pillow-simd
pyvips
selectolax
zipfile-deflate64
